pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "celery>=5.5.3",
    "cryptography>=45.0.7",
    "fastapi>=0.116.1",
    "geoip2>=5.1.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.1.2",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.31.0",
    "spacy>=3.7.6",
    "sqlalchemy>=2.0.43",
    "user-agents>=2.2.0",
    "uvicorn>=0.35.0",
    "email-validator>=2.3.0",
    "sendgrid>=6.12.4",
    "uvloop>=0.21.0",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "httpx>=0.28.1",
]
